fix edgecore cpu usage
